Ajout de l'exercice 2 qui demande un programme de somme de 2 nombres dans le bon format

+++ exercice_2.py
@@ -1,7 +1,7 @@
 
 premier_nombre = input("Bonjour \nVeuillez entrer le premier nombre: ")
 second_nombre = input("entrer le second: ")
-premier_nombre = int(premier_nombre)
-second_nombre = int(second_nombre)
+premier_nombre = float(premier_nombre)
+second_nombre = float(second_nombre)
 somme = premier_nombre + second_nombre
 print("la somme est: ", somme)
